chore: WIP new projectViewItemProcessors plugin (2)
