+++ libs/shared/nx-console-plugins/src/lib/internal-plugins/index.ts
@@ -28,4 +28,5 @@ export const internalPlugins: NxConsolePluginsDefinition = {
     projectNameAndRootStartupMessage,
     nameAndDirectoryStartupMessage,
   ],
+  projectViewItemProcessors: []
 };
